mongo works!

+++ MusShop/migrations/0001_initial.py
@@ -19,7 +19,8 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('count', models.IntegerField()),
-                ('coast', models.DecimalField(decimal_places=2, max_digits=10)),
+                # ('coast', models.DecimalField(decimal_places=2, max_digits=10)),
+                ('coast', models.IntegerField()),
                 ('model', models.CharField(max_length=30)),
             ],
         ),
